Model saving and other changes
Model saving and loading works now.

Support for the 275k image Doom dataset. Dataset code moved to datasets.py.

Training now with continuous amounts of noise, no diffusion timesteps needed.

And as such, buckets for training with certain amounts of noise, so that the noise amounts are sampled as homogenously as possible.